+++ app/lib/publisher/models.dart
@@ -6,9 +6,9 @@ import 'package:gcloud/db.dart' as db;
 
 /// Canonical publisher data.
 @db.Kind(name: 'Publisher', idType: db.IdType.String)
-class Publisher extends db.ExpandoModel {
+class Publisher extends db.ExpandoModel<String> {
   /// The associated domain name of the publisher.
-  String get publisherId => id as String;
+  String get publisherId => id;
 
   /// Markdown formatted description of the publisher.
   ///
@@ -55,9 +55,9 @@ class Publisher extends db.ExpandoModel {
 
 /// Derived publisher data.
 @db.Kind(name: 'PublisherInfo', idType: db.IdType.String)
-class PublisherInfo extends db.ExpandoModel {
+class PublisherInfo extends db.ExpandoModel<String> {
   /// The associated domain name of the publisher.
-  String get publisherId => id as String;
+  String get publisherId => id;
 
   @db.DateTimeProperty()
   DateTime updated;
@@ -88,7 +88,7 @@ abstract class PublisherMemberRole {
 
 /// Stores the membership information of a single user.
 @db.Kind(name: 'PublisherMember', idType: db.IdType.String)
-class PublisherMember extends db.ExpandoModel {
+class PublisherMember extends db.ExpandoModel<String> {
   /// The key of the publisher.
   db.Key get publisherKey => parentKey;