+++ CHANGELOG.md
@@ -3,6 +3,8 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
  * `/documentation/` serving changed: content entry lookup first checks Datastore entity.
+ * Upgraded `gcloud` to `0.7.3`, using the new `delimiter` to recursively
+   delete from storage buckets.
 
 ## `20200610t120907-all`
  * Bumped runtimeVersion to `2020.06.10`.